fix: move scope methods to IdProvider
