+++ src/DIRAC/Resources/IdProvider/IdProvider.py
@@ -26,3 +26,19 @@ class IdProvider:
         """
         self.parameters = parameters
         self.name = parameters.get("ProviderName")
+
+    def getGroupScopes(self, group: str) -> list:
+        """Get group scopes
+
+        :param group: DIRAC group
+        """
+        idPScope = getGroupOption(group, "IdPRole")
+        return scope_to_list(idPScope) if idPScope else []
+
+    def getScopeGroups(self, scope: str) -> list:
+        """Get DIRAC groups related to scope"""
+        groups = []
+        for group in getAllGroups():
+            if (g_scope := self.getGroupScopes(group)) and set(g_scope).issubset(scope_to_list(scope)):
+                groups.append(group)
+        return groups
